organize nearby node searches into folders
